requirements.txt
langchain
ipykernel
python-dotenv
langchain  
langchain-community
langchain_core
faiss-cpu
streamlit
langsmith
langchain-openai
langchain_groq
fastapi
uvicorn
langserve
langchain_chroma
bs4
langchain_huggingface
pypdf
arxiv
wikipedia
duckduckgo-search
PyPDF2 
pymysql
unstructured
validators
pytube
youtube-transcript-api
numexpr
langchain 
tiktoken

